fix merge mistake (reintroduced an old warning)

+++ Fileinfo.cc
@@ -80,7 +80,8 @@ Fileinfo::fillwithbytes(enum readtobuffermode filltype,
     char buffer[4096];
     while (f1) {
       f1.read(buffer, sizeof(buffer));
-      chk.update(f1.gcount(), buffer);
+      // gcount is never negative, the cast is safe.
+      chk.update(static_cast<std::size_t>(f1.gcount()), buffer);
     }
 
     // store the result of the checksum calculation in somebytes
